# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Minimarket Plus API",
        "description": "API REST del backend de Minimarket Plus: productos, carrito, inventario, ventas y usuarios. Documentada con OpenAPI e hipermedios HATEOAS. Los recursos est\u00e1n protegidos por roles (CLIENTE, CAJERO, ADMIN).",
        "contact": {
            "name": "Equipo Minimarket Plus"
        },
        "license": {
            "name": "Uso acad\u00e9mico - Duoc UC"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8084",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/carrito/{id}": {
            "get": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Obtener \u00edtem del carrito por ID",
                "operationId": "obtenerCarritoPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "\u00cdtem no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "\u00cdtem encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Actualizar \u00edtem del carrito",
                "operationId": "actualizarCarrito",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrito"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "\u00cdtem no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "\u00cdtem actualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Eliminar producto del carrito",
                "operationId": "eliminarProductoDelCarrito",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "\u00cdtem no encontrado"
                    },
                    "204": {
                        "description": "Producto eliminado del carrito"
                    }
                }
            }
        },
...
        "/api/carrito": {
            "get": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Listar carrito",
                "description": "Devuelve los productos en el carrito con enlaces HATEOAS.",
                "operationId": "listarCarrito",
                "responses": {
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelCarrito"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollectionModelEntityModelCarrito"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Agregar producto al carrito",
                "description": "A\u00f1ade un producto al carrito validando disponibilidad de stock.",
                "operationId": "agregarProductoAlCarrito",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrito"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "403": {
                        "description": "Sin permisos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Producto agregado al carrito",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Rol": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "roles": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rol"
                        }
                    }
                }
            },
...
            "Links": {
                "type": "object",
                "additionalProperties": {
                    "$ref": "#/components/schemas/Link"
                }
            },
            "Categoria": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Producto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoria": {
                        "$ref": "#/components/schemas/Categoria"
                    }
                }
            },
...
            "Carrito": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "EntityModelCarrito": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "producto": {
                        "$ref": "#/components/schemas/Producto"
                    },
                    "cantidad": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "CollectionModelEntityModelCarrito": {
                "type": "object",
                "properties": {
                    "_embedded": {
                        "type": "object",
                        "properties": {
                            "carritoList": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/EntityModelCarrito"
                                }
                            }
                        }
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
            "Link": {
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "hreflang": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "deprecation": {
                        "type": "string"
                    },
                    "profile": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "templated": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}